Disable HttpClient automatic retries
We want to disable the automatic retries mechanism brought by HttpClient and use instead the internal retry that has been written.
That way we will first be less verbose even when we set `quiet` to `true` but we will make sure we are using the code that has been written to do the checks.